+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mining/Mining.java
@@ -2,7 +2,7 @@ package insane96mcp.iguanatweaksreborn.module.mining;
 
 import insane96mcp.iguanatweaksreborn.module.mining.feature.CustomHardness;
 import insane96mcp.iguanatweaksreborn.module.mining.feature.GlobalHardness;
-import insane96mcp.iguanatweaksreborn.module.mining.feature.InstaMineSilverfish;
+import insane96mcp.iguanatweaksreborn.module.mining.feature.MiningMisc;
 import insane96mcp.iguanatweaksreborn.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -12,14 +12,14 @@ public class Mining extends Module {
 
 	public GlobalHardness globalHardness;
 	public CustomHardness customHardness;
-	public InstaMineSilverfish instaMineSilverfish;
+	public MiningMisc miningMisc;
 
 	public Mining() {
 		super(Config.builder);
 		pushConfig(Config.builder);
 		globalHardness = new GlobalHardness(this);
 		customHardness = new CustomHardness(this);
-		instaMineSilverfish = new InstaMineSilverfish(this);
+		miningMisc = new MiningMisc(this);
 		Config.builder.pop();
 	}
 
@@ -28,6 +28,6 @@ public class Mining extends Module {
 		super.loadConfig();
 		globalHardness.loadConfig();
 		customHardness.loadConfig();
-		instaMineSilverfish.loadConfig();
+		miningMisc.loadConfig();
 	}
 }